+++ js-miniapp-sdk/src/miniapp.ts
@@ -1,11 +1,13 @@
 import { InterstitialAdResponse } from './types/responseTypes/interstitial';
 import { MiniAppPermissionType } from './MiniAppPermissionType';
+import { RewardedAdResponse } from './types/responseTypes/rewarded';
 /**
  * A module layer for webapps and mobile native interaction.
  */
-interface MiniApp {
+interface MiniAppFeatures {
   /** @returns The Promise of provided id of mini app from injected side. */
   getUniqueId(): Promise<string>;
+
   /** @returns The Promise of permission result of mini app from injected side. */
   requestLocationPermission(): Promise<string>;
 }
@@ -14,13 +16,44 @@ interface MiniApp {
  * A contract declaring the interaction mechanism between mini-apps and native host app to display ads.
  */
 interface Ad {
-  /** @returns The Promise of interstitial ad response result from injected side. */
-  showInterstitialAd(): Promise<InterstitialAdResponse>;
+  /**
+   * Loads the specified Interstittial Ad Unit ID.
+   * Can be called multiple times to pre-load multiple ads.
+   * Promise is resolved when successfully loaded.
+   * @returns The Promise of load ad response result from injected side.
+   * Promise is rejected if failed to load.
+   */
+  loadInterstitialAd(id: string): Promise<null | Error>;
+
+  /**
+   * Loads the specified Rewarded Ad Unit ID.
+   * Can be called multiple times to pre-load multiple ads.
+   * Promise is resolved when successfully loaded.
+   * * @returns The Promise of load ad response result from injected side.
+   * Promise is rejected if failed to load.
+   */
+  loadRewardedAd(id: string): Promise<null | Error>;
+
+  /**
+   * Shows the Interstitial Ad for the specified ID.
+   * Promise is resolved after the user closes the Ad.
+   * @returns The Promise of Interstitial ad response result from injected side.
+   * Promise is rejected if the Ad failed to display wasn't loaded first using MiniApp.loadRewardedAds.
+   */
+  showInterstitialAd(id: string): Promise<InterstitialAdResponse>;
+
+  /**
+   * Shows the Rewarded Ad for the specified ID.
+   * Promise is resolved with an object after the user closes the Ad. The object contains the reward earned by the user.
+   * Reward will be null if the user did not earn the reward.
+   * @returns The Promise of Rewarded ad response result from injected side.
+   * Promise is rejected if the Ad failed to display wasn't loaded first using MiniApp.loadRewardedAds.
+   */
+  showRewardedAd(id: string): Promise<RewardedAdResponse>;
 }
 
-/** @internal */
 /* tslint:disable:no-any */
-export class MiniAppImp implements MiniApp, Ad {
+export class MiniApp implements MiniAppFeatures, Ad {
   private requestPermission(permissionType: string): Promise<string> {
     return (window as any).MiniAppBridge.requestPermission(permissionType);
   }
@@ -33,7 +66,19 @@ export class MiniAppImp implements MiniApp, Ad {
     return this.requestPermission(MiniAppPermissionType.LOCATION);
   }
 
-  showInterstitialAd(): Promise<InterstitialAdResponse> {
-    return (window as any).MiniAppBridge.showInterstitialAd();
+  loadInterstitialAd(id: string): Promise<null | Error> {
+    return (window as any).MiniAppBridge.loadInterstitialAd(id);
+  }
+
+  loadRewardedAd(id: string): Promise<null | Error> {
+    return (window as any).MiniAppBridge.loadRewardedAd(id);
+  }
+
+  showInterstitialAd(id: string): Promise<InterstitialAdResponse> {
+    return (window as any).MiniAppBridge.showInterstitialAd(id);
+  }
+
+  showRewardedAd(id: string): Promise<RewardedAdResponse> {
+    return (window as any).MiniAppBridge.showRewardedAd(id);
   }
 }