+++ js-miniapp-sdk/src/types/adTypes/index.ts
@@ -1,5 +1,6 @@
 /**
- * Supported ad formats
+ * Enum for supported ad types/formats
+ * These ad formats replicate formats supported by Google
  */
 export enum AdTypes {
   INTERSTITIAL,