+++ js-miniapp-sdk/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### v1.2.0 (In progress)
 
-- Added support for requesting show interstitial ad from the host application to allow displaying Interstitial ads [See here](README.MD#4-Show-Interstitial-Ad)
+- Added support for requesting the load and display of Interstitial & Rewarded Ads by the host application [See here](README.MD#4-Show-Ads)
 
 ### v1.1.0 (2020-7-21)